Remove unnecessary casts

+++ src/datasettingsdialog.cpp
@@ -20,23 +20,23 @@ DataSettingsDialog::DataSettingsDialog(QWidget *parent) : QDialog{ parent }
     auto settingsLayout{ new QGridLayout{ settingsBox } };
 
     auto dataFormatLabel{ new QLabel{ u"Data format"_s } };
-    dataTypeBox->addItem(u"ASCII"_s, static_cast<int>(SerialTransceiver::DataTypes::ascii));
-    dataTypeBox->addItem(u"unsigned 8 bit"_s, static_cast<int>(SerialTransceiver::DataTypes::u8));
-    dataTypeBox->addItem(u"unsigned 16 bit"_s, static_cast<int>(SerialTransceiver::DataTypes::u16));
-    dataTypeBox->addItem(u"unsigned 32 bit"_s, static_cast<int>(SerialTransceiver::DataTypes::u32));
-    dataTypeBox->addItem(u"unsigned 64 bit"_s, static_cast<int>(SerialTransceiver::DataTypes::u64));
-    dataTypeBox->addItem(u"signed 8 bit"_s, static_cast<int>(SerialTransceiver::DataTypes::i8));
-    dataTypeBox->addItem(u"signed 16 bit"_s, static_cast<int>(SerialTransceiver::DataTypes::i16));
-    dataTypeBox->addItem(u"signed 32 bit"_s, static_cast<int>(SerialTransceiver::DataTypes::i32));
-    dataTypeBox->addItem(u"signed 64 bit"_s, static_cast<int>(SerialTransceiver::DataTypes::i64));
-    dataTypeBox->addItem(u"float 32 bit"_s, static_cast<int>(SerialTransceiver::DataTypes::f32));
-    dataTypeBox->addItem(u"double 64 bit"_s, static_cast<int>(SerialTransceiver::DataTypes::f64));
+    dataTypeBox->addItem(u"ASCII"_s, SerialTransceiver::DataTypes::ascii);
+    dataTypeBox->addItem(u"unsigned 8 bit"_s, SerialTransceiver::DataTypes::u8);
+    dataTypeBox->addItem(u"unsigned 16 bit"_s, SerialTransceiver::DataTypes::u16);
+    dataTypeBox->addItem(u"unsigned 32 bit"_s, SerialTransceiver::DataTypes::u32);
+    dataTypeBox->addItem(u"unsigned 64 bit"_s, SerialTransceiver::DataTypes::u64);
+    dataTypeBox->addItem(u"signed 8 bit"_s, SerialTransceiver::DataTypes::i8);
+    dataTypeBox->addItem(u"signed 16 bit"_s, SerialTransceiver::DataTypes::i16);
+    dataTypeBox->addItem(u"signed 32 bit"_s, SerialTransceiver::DataTypes::i32);
+    dataTypeBox->addItem(u"signed 64 bit"_s, SerialTransceiver::DataTypes::i64);
+    dataTypeBox->addItem(u"float 32 bit"_s, SerialTransceiver::DataTypes::f32);
+    dataTypeBox->addItem(u"double 64 bit"_s, SerialTransceiver::DataTypes::f64);
     settingsLayout->addWidget(dataFormatLabel, 0, 0);
     settingsLayout->addWidget(dataTypeBox, 0, 1);
 
     auto byteOrderLabel{ new QLabel{ u"Byte order (for numeric)"_s } };
-    byteOrderBox->addItem(u"little-endian"_s, static_cast<int>(QDataStream::LittleEndian));
-    byteOrderBox->addItem(u"big-endian"_s, static_cast<int>(QDataStream::BigEndian));
+    byteOrderBox->addItem(u"little-endian"_s, QDataStream::LittleEndian);
+    byteOrderBox->addItem(u"big-endian"_s, QDataStream::BigEndian);
     settingsLayout->addWidget(byteOrderLabel, 1, 0);
     settingsLayout->addWidget(byteOrderBox, 1, 1);
 
@@ -86,14 +86,13 @@ void DataSettingsDialog::loadSettings()
     QSettings settingsFile;
     settingsFile.beginGroup(u"data"_s);
     settings.dataType = static_cast<SerialTransceiver::DataTypes>(
-            settingsFile.value(u"type"_s, static_cast<int>(SerialTransceiver::DataTypes::ascii))
-                    .toInt());
+            settingsFile.value(u"type"_s, SerialTransceiver::DataTypes::ascii).toInt());
     settings.byteOrder = static_cast<QDataStream::ByteOrder>(
             settingsFile.value(u"byteOrder"_s, QDataStream::LittleEndian).toInt());
     settingsFile.endGroup();
 
-    updateIndex(dataTypeBox, static_cast<int>(settings.dataType));
-    updateIndex(byteOrderBox, static_cast<int>(settings.byteOrder));
+    updateIndex(dataTypeBox, settings.dataType);
+    updateIndex(byteOrderBox, settings.byteOrder);
 }
 
 /**
@@ -105,7 +104,7 @@ void DataSettingsDialog::saveSettings()
     QSettings settingsFile;
 
     settingsFile.beginGroup(u"data"_s);
-    settingsFile.setValue(u"type"_s, static_cast<int>(settings.dataType));
+    settingsFile.setValue(u"type"_s, settings.dataType);
     settingsFile.setValue(u"byteOrder"_s, settings.byteOrder);
     settingsFile.endGroup();
 }
